+++ src/App.tsx
@@ -2,14 +2,14 @@ import { GlobalStyle } from '@/styles/global'
 import { ThemeProvider } from '@emotion/react'
 import { theme } from '@/styles/theme'
 
-import MainLayout from '@/components/MainLayout'
+import Main from '@/components/Main'
 
 function App() {
   return (
     <>
       <ThemeProvider theme={theme}>
         <GlobalStyle />
-        <MainLayout />
+        <Main />
       </ThemeProvider>
     </>
   )
