+++ src/components/Main/AutoRollingBar/NewsBar.tsx
