+++ src/components/Main/AutoRollingBar/index.tsx
