+++ src/components/Main/Media/ListView/NewsCard/NewsContent/Main/Thumbnail.tsx
