+++ src/components/Main/Media/ListView/NewsCard/NewsContent/Main/index.tsx
