+++ src/components/Main/Media/ListView/NewsCard/NewsContent/Sub/index.tsx
