+++ src/components/Main/Media/ListView/NewsCard/NewsContent/index.tsx
