+++ src/components/Main/Media/ListView/NewsCard/NewsInfo/BrandMark.tsx
