+++ src/components/Main/Media/ListView/NewsCard/NewsInfo/index.tsx
