+++ src/components/Main/Media/ListView/NewsCard/index.tsx
